Write a C++ program for insert in the first position of linked list.
First create and linked list.
• For inserting we have to ensure a new linked list in dynamically and assign the value
of data part. We can set the value by default or by user input.
• To add in the first position we have to change the head first and our previous address
part sets in the new address section to link the nodes.
• And our new head is p (which is the new node).
• Then print the data by using the while loop we need to traverse all the elements and
showing the data by using a while loop. When the head is not NULL then loop going
forward and print the next data. Gradually it will print all the data and when it hits
NULL our loop will stop.