Write a C++ program for insert in the last position of linked list.
First create and linked list.
• For inserting we have to ensure a new linked list in dynamically and assign the value
of data part. We can set the value by default or by user input.
• Set the first node = head. And a temporary variable temp= head. Traverse all the
elements to reach the last position by (temp -> next != NULL). Then the address part
of p is NULL and put the address of p to the last link list.
• Then print the data by using the while loop we need to traverse all the elements and
showing the data by using a while loop. When the head is not NULL then loop going
forward and print the next data. Gradually it will print all the data and when it hits
NULL our loop will stop.