Write a C++ program for insert in the specific position of linked list.
First create and linked list.
• For inserting we have to ensure a new linked list in dynamically and assign the value
of data part. We can set the value by default or by user input.
• Here p is the new node and we set the node after b. For this insert we have to change
the address of c into the p and the address of the p set to b.
• Then print the data by using the while loop we need to traverse all the elements and
showing the data by using a while loop. When the head is not NULL then loop going
forward and print the next data. Gradually it will print all the data and when it hits
NULL our loop will stop.