Write a C++ program for delete in the last position of linked list.
• First create and linked list.
• Declare a node prev which is equal to a. Then we traverse the prev through next to next
which is not equal to zero and it will pass one to another.
• Then we evaluate zero to the address part of prev and delete this.
• Now we declare our head again and print the data by using the while loop we need to
traverse all the elements and showing the data by using a while loop. When the head is
not NULL then loop going forward and print the next data. Gradually it will print all
the data and when it hits NULL our loop will stop.