Write a C++ program for delete in the specific position of linked list.
• First create and linked list. Set the value to delete.
• We declare two node types variable current and prev (which is NULL).
• We set a key value to delete then we declare our current to head. We use a while loop
for when the current is not equal to zero then our loop will start and if the key value is
equal to current data value then our prev->next will be current->next , and delete the
current. Then the previous will be current now.
• Then change to head to temp and print the data by using the while loop we need to
traverse all the elements and showing the data by using a while loop. When the head is
not NULL then loop going forward and print the next data. Gradually it will print all
the data and when it hits NULL our loop will stop.